+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Fixed:
  - Doesn't throw exception on http errors (4xx and 5xx statuses)
 
+### Added:
+ - Metro, Auth
+
 ## v0.2.6 - 2017-05-05
 ### Changed:
  - Tests and refactoring